consolidate progression service
