+++ services/progression_service.py
@@ -6,6 +6,8 @@
 
 import logging
 import time
+from threading import Lock
+from typing import Dict, Set
 from services.modifiers_utils import (
     parse_json_field,
     merge_modifiers_with_rules,
@@ -41,6 +43,92 @@ from services.modifiers_utils import (
 )
 
 logger = logging.getLogger(__name__)
+
+# --------------------------------------------------------
+# Simplified In-Memory Progression State (for tests)
+# --------------------------------------------------------
+
+# Lock protecting state mutations when used in async tests
+_state_lock = Lock()
+
+# Castle level tracked in memory
+castle_state: Dict[str, int] = {"level": 1}
+
+# Sets/dicts for nobles and knights
+nobles: Set[str] = set()
+knights: Dict[str, Dict[str, int]] = {}
+
+
+def progress_castle() -> int:
+    """Increase the castle level by one."""
+    with _state_lock:
+        new_level = castle_state.get("level", 1) + 1
+        castle_state["level"] = new_level
+        logger.info("🏰 Castle leveled up! New level: %s", new_level)
+        return new_level
+
+
+def add_noble(name: str) -> None:
+    """Add a noble if not already present."""
+    with _state_lock:
+        added = name not in nobles
+        nobles.add(name)
+        if added:
+            logger.info("👑 Noble '%s' added.", name)
+        else:
+            logger.debug("Noble '%s' already exists.", name)
+
+
+def remove_noble(name: str) -> None:
+    """Remove a noble if they exist."""
+    with _state_lock:
+        existed = name in nobles
+        nobles.discard(name)
+        if existed:
+            logger.info("❌ Noble '%s' removed.", name)
+        else:
+            logger.debug("Noble '%s' not found.", name)
+
+
+def add_knight(knight_id: str, rank: int = 1) -> None:
+    """Register a new knight."""
+    with _state_lock:
+        if knight_id in knights:
+            logger.debug("Knight '%s' already exists.", knight_id)
+            return
+        knights[knight_id] = {"rank": rank}
+        logger.info("🛡️ Knight '%s' added with rank %s.", knight_id, rank)
+
+
+def promote_knight(knight_id: str) -> int:
+    """Promote a knight to the next rank and return the new rank."""
+    with _state_lock:
+        knight = knights.get(knight_id)
+        if not knight:
+            logger.error("⚠️ Cannot promote: Knight '%s' not found.", knight_id)
+            raise ValueError("Knight not found")
+        knight["rank"] += 1
+        new_rank = knight["rank"]
+        logger.info("⬆️ Knight '%s' promoted to rank %s.", knight_id, new_rank)
+        return new_rank
+
+
+def get_state() -> Dict[str, object]:
+    """Return a snapshot of the current progression state."""
+    with _state_lock:
+        return {
+            "castle": castle_state.copy(),
+            "nobles": set(nobles),
+            "knights": {k: v.copy() for k, v in knights.items()},
+        }
+
+
+def reset_state() -> None:
+    """Reset all in-memory progression data."""
+    with _state_lock:
+        castle_state.update({"level": 1})
+        nobles.clear()
+        knights.clear()
 
 # Number of seconds before cached modifiers expire.
 _CACHE_TTL = 60
